update Cordova demo with background processing
and update README.md

Note that the background processing enhancements are quite
straightforward for native platform code on Android & iOS.

It would have been possible to include the background processing
enhancements in the `object-batch-connection` classes.

However, keeping the background processing enhancements in the
Cordova plugin code makes it easier to fine tune if needed
for Cordova plugin enhancements and React Native support.