rename & refactor ->SQLite object batch connection
update title to: SQLite object batch connection core

factor out SQLiteBatchCore classes for Android & iOS

that can be directly used in a plugin on Cordova &
potentially React Native

handle all numbers as doubles in SQLiteBatchCore classes
(with no 64-bit integer support) to maintain consistency
on all supported platforms on Apache Cordova & React Native

update the native Cordova demo plugin object class name

apply minor cleanup of imports on Android in the Cordova demo plugin

remove outdated comments from the Java code

update the documentation in README.md

Notes:

The SQLiteBatchCore API is a synchronous API on Android.

Background processing can easily be implemented in the
component-specific code for Apache Cordova or React Native,
if desired.

It is intended to keep the SQLiteBatchCore classes for Android & iOS
in this repository in order to preserve the development history
of this code.

The low-level C and Java classes may be moved out to a separate
repository at some point in the future.